Updated overzicht page
Added code to get the last used counter for the kenmerk. Which is passed to the database.
Moved the recordsPerPage and page variables to overzicht.php.
Added code to get the corresponding document type for the post gotten from the database.
Filled in the remainder of the variables for the overzicht page.
The post ID is now displayed as a link to the postin page in order to edit the post.
Added code to navigate to the previous and next page in the overzicht page.
Added setup to display the page numbers on the bottom of the overzicht.