Update for overzicht and displaying post when selected from overzicht
Added a function to update the data in the database with the given ID from the post.
Changed the table name REGISTRY to post.
Added code to postin.php and postuit.php to get the id of the post being looked up in order to get the right information from the database to display in the forms.
The render function now gives the post needed as an array to the form and a submitValue to fill in into the button on the bottom of the page.
Both forms now automatically fill in data from the selected post if chosen from the overzicht, otherwise the default or empty values are shown.
Removed javascript code that is no longer needed for twig is being used.
Updated the upload_post_(in/out).php files to check for minimum of 12 values given (still need to check that better) and then checking whether the data is to be updated or uploaded in the database. Otherwise it will display a warning. (still not solid though)