Added autocomplete functionality to the forms
Also updated the way the kenmerk is retrieved for a new post. Didn't work properly.
Added php file to get the information from the database.
Adjusted the design.html file for the autcomplete functionality of jquery, was conflicting with the normal jquery.
Javascript code to make the autocomplete work in the forms.