Added functionality to go back to the page where the aciton was done
This for when navigated to the post in or post uit page from the zoeken of geavanceerd zoeken page.
Otherwise the user would be sent back to the same page, which makes it less user friendly.
Also check if everything is filled in correctly, as in not being empty. Giving a warning if it is empty.
Furthermore removed the action parameters in the forms so it points to itself, where the checks happen.