Updated uploaded documents buttons
Changed the type of button from input - button to just button with a form-action variable. Not sure if the right solution right now.
Added css code to make sure the text doesn't go out of bounds.
Added images on the buttons to display the actions as an image instead of text. Although not yet sure which image to use for the "open" button.