Changed the way files are deleted from the post
Added check to see if the file exists before removing the file.
When clicking the delete button, no longer an form action is done, but an ajax call is being made thus preventing the entered data from being lost before saving it.
Removed the id's of the buttons and replaced it with classes so the exact buton can be checked when clicked. Before it would just check the top most button list item's button.
When the call is made to delete the button, and it is succesfull, the button's are removed and the tag "(removed)" is placed in the innertext of the list item.