Update to save and show number of files per post
Added variable to the post class including a getter.
Updated the posts class to save the number of files correctly on both upload and update.
Redesigned the table on the search forms. Adding the number of files column, color of post without number of files will be set to yellow.