Changes to autofill information in the post in and uit form
Added parameter our_loginname to class post.
Added our_loginname to the queries that upload and edit the post when the post in or post uit form is submitted.
Added php file to handle the ajax call for autofilling the information.
Added ajax call to both post in and post uit and adjusted the corresponding fields. Furthermore added a hidden input field for the loginname to be submitted.
And lastly did some cleanup on the code/application.