Fixed the saving of the number of files uploaded
Adjusted the function uploadPost to make sure the files are firstly written/saved in the directory before inserting the record in the database. In this way the count of files is done by checking the amount of files that are saved.
Furthermore left some code in commentary to show what has been tried to check if the variable $_FILES is "empty" or not.