Added check for ownership and rights
Changed the code so when the user has no ownership of the uploaded post or is no admin or functional manager he or she has no rights to adjust the data.
Did the above also for when users try to update a post by hacking their way into the html. So the data doesn't get saved when by change an update is made by hacking their way in.
Fields that should not be displayed are not displayed or disabled.
Furthermore adjusted the layout of the code slightly to make it better looking.